requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
slack-bolt==1.18.0
slack-sdk==3.27.0
anthropic>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0
rich>=13.0.0
click>=8.1.0
python-dotenv>=1.0.0
